adding docs to HadoopFormat and moving it to its own file
